virtio-video: create virtio_devices for decoder and encoder respectively
Create and register two virtio_devices for encoder and decoder respectively
so as to trigger probe() call for encoder and decoder separately.

Hab/vhost creates only one virtio_device, which was shared by both decoder
and encoder. Now the device is made as the parent device of the two
virtio_devices created in this module.

Change-Id: Ie8b9f5d9290616e0bcd6892b4fb3997788f9686e