virtio-video: use HAB messages to exchange feature bits and configuration space
1. Open hab and get feature bits and configuration space from BE.
Implement it in virtio_device_reset function.
2. Implement feature finalization. Currently, it only loops back
whatever FE gets from BE. In the future, MSM_HAB bit will be
added to inform FE that qualcomm's backend and qualcomm's extra
feature bits are enabled.

Background:

According to MMIO-based virtio specification, virtio device is
required to provide a set of memory-mapped registers from BE to
FE. The registers include device version, feature bits, and device
configuration space. However, HAB has not implemented those
facilities yet. Virtio-video, as well as other virtio devices,
is thus forced to hard-code feature bits and configurations in FE.

To be better aligned with upstream, we need to be able to generate
and exchange feature bits and device configuration space dynamically.
For example, max_caps_length tells FE the max length of the response
of the capabilities query; a new feature bit MSM_HAB will be
added to enable FE to confirm if qualcomm's backend and qualcomm's
extra features are enabled.

This pair of gerrits are to workaround the HAB limitation by
exchanging a memory buffer back-and-forth. This will facilitate the
incoming gerrits to further align with the existing upstream
virtio-video driver.

Change-Id: Ia1d52906d56194f52dbd4f51531d276f59895413