virtio-video: depreciate done_buffer in the stream struct
The done_buffer was used to pair the data buffer and its
corresponding meta buffer in handling of the EBD/FBD events.
The pairing made it was possbile to make two vb2_buffer_done() calls
into one atomic operation with a mutex.

Given we had removed the mutex for performace improvement and
removed meta buffer poll_wait in previous commits, the use of
done_buffer is redundent.

Change-Id: Ie5d6a1767d8fe67f6daad690642cefe942822419