virtio-video: trigger EOS event after receiving meta buffer
When Codec2 receives EOS event, it will dequeue both data
and meta buffer. Previously, EOS event is triggered after receiving
data buffer, therefore there is a chance that meta buffer is not
ready when Codec2 tries to dequeue it.

To solve this issue, add EOS flags for meta buffer in BE as well and
trigger EOS event only after receiving meta buffer.

Change-Id: Ib528740e57819044750c544278640b31819565e3